### 2024-10-08 #7330 @editorialbot
[REVIEW]: FACILE-RS: archiving and long term preservation of research software repositories made easy
**Submitting author:** @MarieHouillon (<a href="http://orcid.org/0000-0002-6584-0233">Marie Houillon</a>)
**Repository:** https://github.com/openCARP-org/FACILE-RS
**Branch with paper.md** (empty if default branch): joss
**Version:** v3.3.0
**Archive:** 10.35097/g2jahr9cj5ckf3qr
