### 2024-10-16 #7330 @exaexa
Hello @MarieHouillon, I'm writing some comments to the points in the review.

In general, FACILE-RS is a very valid effort relevant and potentially useful to all authors of open research software. Reliable catalogization and registration of research software in all these repositories and registries around is certainly an issue, and while everyone agrees that it should be "just done", the developers (including me) may get quite annoyed when facing the amount of petty tasks and clicking required to get everything to a perfect state.

FACILE-RS, if installed and configured properly, removes a substantial part of this burden from researchers.

In summary, I have 1 major comment about the documentation structure (see below). Other comments are minor and might not need be addressed, but it would be great to at least discuss these to see if FACILE-RS can be made more useful for the RSD crowd.

### Scope and scholarly effort

FACILE-RS satisfies the exact definition of "Research software" as [given by JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) only as a borderline match (technically it supports functioning of research instruments, for some very general definition of research instruments). On the other hand, given the utility of the software to all JOSS readers and authors, I'd still say it's in scope.

The coding effort behind the software matches the JOSS expectations, moreover the code was used previously for a few other research packages. Lines of code are not a very good metric for guessing the effort; the main value of this package is that the CI pipelines are tested and working, which takes quite a lot of effort. I can imagine my colleagues would cite this in papers about larger-scale software and data FAIRification; I would certainly use&cite this for any future configuration- and version-sensitive research software.

### Functionality

(I still have to verify this on my setup, but given the existing deployments I assume the functionality is already proven to be OK. Some comments provided in *Documentation* section also apply to the installation procedure.)

### Documentation

As the only major comment from my side, the documentation of FACILE-RS would deserve some revisions in order to really satisfy the expectations. In overview, although the individual parts are documented quite well, the "integration" view is largely missing.

- **Installation**: The README should ideally provide the "deal" view of what users need to do in order to receive which functionality. In particular, it should clearly state what the "cost" for the user is (creating the CodeMeta files and modifying the GitLab CI?) to receive the "benefit" (these are documented well). Some extra comments:
  - The installation instructions should be as specific and unambiguous as possible; which is not the case e.g. right in the first instruction. E.g., https://git.opencarp.org/openCARP/FACILE-RS#setup says "add the following to your job", without specifying which exact job (there might be some where this simply does not work), without providing a reference to what this even means for potential users who are new to GitLab CI, or without clarifying the implications of including other people's code in your CI.
  - The follow-up paragraph about actual installation ("Adapting from FACILE-RS") should have a recipe structure, preferably with the individual tasks listed as steps and substantiated (why do I need `PUSH_TOKEN` as well as `PRIVATE_TOKEN`?)
  - Defaults should be "safe" -- preferably the software should not do any externally visible actions (such as pushing to RADAR) without getting an explicit instruction to do so by the users.
  - It is not quite clear how the upgrade process is handled with this installation method. In particular, the `pip install` should point to a tagged stable version (or ideally "a major semver release"), which would allows people to prevent impact of future breaking changes. (Compare with e.g. GitHub actions versioning.)

  I'd whole-heartedly recommend addressing the issues with the installation documentation by creating an explicit "template repository" where all CI is set up and all external actions are disabled by default. With that, people can simply clone the template and merge it to their projects using git. Notably, such a template repository would also provide a quite sustainable configuration upgrade management workflow (users may just repeatedly merge the updates from the template and solve possible conflicts).
- **Example use**: The FACILE-RS repository serves as an example, which is OK but not quite as useful as it could be (notably, the example should preferably not be specific to any real software package, to prevent user confusion). See above for a possible improvement.
- **Automated tests**: Same as above. The CI kindof "tests itself" by "working right" in the repository, but having a separate repository with a "cleaned" mock example project would be much better. I don't see any actual test suite, but that might not be necessary in this case.
- **Community**: A bit of documentation space should be spent on documenting the extension process. In particular, it is now unclear to me how complex it is to:
  - extend FACILE-RS to also send information to different registries and repositories (e.g., registering versions in https://bio.tools/ , automated Fediverse announcements, ...)
  - run the FACILE-RS scriptage on GitHub (despite all openness issues of the proprietary platform, GitHub will probably continue to be a major software publishing platform in the foreseeable future)

Random extra notes:
- The documentation is inconveniently scattered over several places -- some of the scripts are documented in the README as well as in the generated docs at readthedocs.io (is there a difference between these two?). The tutorials are available as markdown in a subdirectory, while they would fit much better into the readthedocs site to enable cross-linking of the functions.
- Some API/tool names should preferably be adjusted to exactly refer to what the scripts do. For example, `create_` prefix is used for generating data from the CodeMeta (assumingly!), sometimes it means packing a package, and sometimes it also means uploading the package. `prepare_` means either modifying metadata fields, or registering a DOI online. the `run_*_pipeline` scripts are GravCMS specific without specifying that in the name (what if users want to extend this with another CMS support?). To alleviate the issue, it is better to use precise verbs for the command names (`generate_`, `register_`, `update_`, `pack_`, `upload_`, `pack_and_upload_`, ...). On a side thought, to prevent future name clashes with other tools, it is beneficial to give a common prefix to all command line utility names (e.g., `facilers_update_release_metadata` ? Or just `frs_` ?).

### Paper

- **State of the field**: The paper mentions some related metadata standards and repositories but no clear statement is provided on whether there is a work-alike software, except perhaps for the implicit mention of HERMES project. Explicitly referring to other software (or at least to Table 2 in the HERMES paper, if still up-to-date/relevant) might be more useful for readers. If FACILE-RS is unique in this regard (or in some properties), this should be highlighted too.
- **Quality of writing**: I don't see any obvious issues, perhaps except for the several "clickable-only" links to websites (which will eventually break). These should be ideally replaced by citations, or perhaps avoided altogether (among other, the links to GitLab CI/CD docs are not required, and clickable link to tutorials on L81 is already implied by context).
- On a side note, it is a little confusing that the tools in Table 1 have different identifiers than what is seen in Figure1. Is the exact identification of job steps required for the readers in the paper? Perhaps visualizing the connection of data sources and targets would provide a better illustration for the readers? (I could imagine something like an arrow going from CodeMeta file through a given job step to Radar repository, but other ways might be much better.)
