### 2024-11-21 c272529
Update paper with Zenodo integration + CLI
