### 2024-11-22 b174f8d
Add --assets-token and --assets-token-name arguments
