+++ facile_rs/utils/http.py
@@ -10,7 +10,7 @@ import yaml
 logger = logging.getLogger(__file__)
 
 
-def fetch_files(locations, path):
+def fetch_files(locations, path, headers={}):
     """Fetch files from local locations or a URLs and save them at the given path.
 
     :param locations: list of URL or paths to the files
@@ -22,9 +22,8 @@ def fetch_files(locations, path):
         target = path / location.split('/')[-1]
 
         logger.debug('location = %s, target = %s', location, target)
-
         if urlparse(location).scheme:
-            response = requests.get(location)
+            response = requests.get(location, headers=headers)
             response.raise_for_status()
             with open(target, 'wb') as f:
                 f.write(response.content)
