### 2024-11-25 #7330 @exaexa
@MarieHouillon good evening! I'll try to get to re-evaluating as soon as possible. I was still unable to find a moment to try the functionality myself, but I hope I'll get to it this week. From your description&quick look I can assume that everything should be OK from my side -- the changes are indeed comprehensive, and the docs + template + scriptage style now look super accesible!
