### 2024-11-25 #7330 @exaexa
Review notes after the revision:
- All issues in Documentation seem solved to me now. Specifically for the "community guidelines", the small note about other git hosting platforms + centralization of the scripts (which removes almost all guesswork from the extension process) seems sufficient to me.
- Reproducibility &c.: This is a TODO from me (will edit here later and check all the other boxes)
- I spent some time thinking about the scope eligibility and decided that this fits into the definition, so I'm ticking the box there too.

-----------------

As a note for @atrisovic and/or other editors: It might be useful to make it a bit clearer in JOSS documentation whether the "meta" research software (i.e., one that supports other research software as a primary goal) is in the JOSS scope. My humble opinion is that it should be, but better be sure.

In this particular case, I didn't manage to find a JOSS-published software that would serve as a precedens for deciding directly. But I can argue by comparison that Facile-RS would belong to the same category as various linear algebra libraries and workflow execution managers -- these technically don't do any research-related goals by themselves, only help assembling&running the actual research code, yet they still clearly fall into the "research software" scope. Another way to argue is to categorize Facile-RS as a curation tools in the "librarian" science, belonging to the contents of [LD4](https://sites.google.com/view/2024-ld4-conference/home) or [swiss repro](https://www.reproducibility.ch/) or similar conferences, thus fitting the scope.
