### 2024-12-12 172755c
Test install on all supported Python versions
