### 2024-12-13 7739202
Add CI job for coverage and optional dependencies for testing
