### 2024-12-13 65af0d9
Test create_bag and create_bagpack
