### 2024-12-13 5e5d531
Add test for standard output for cff and datacite scripts
