+++ facile_rs/tests/test_create_datacite.py
@@ -39,3 +39,17 @@ def test_env(monkeypatch, tmpdir):
     main()
     with open(path.join(SCRIPT_DIR, 'datacite_ref.xml')) as datacite_ref:
         assert output_datacite.read() == datacite_ref.read()
+
+def test_stdout(monkeypatch, capsys):
+    monkeypatch.setattr('sys.argv',
+                        [
+                            sys.argv[0],
+                            '--codemeta-location', CODEMETA_LOCATION,
+                            '--creators-location', CREATORS_LOCATIONS,
+                            '--contributors-location', CONTRIBUTORS_LOCATIONS,
+                        ])
+    main()
+    captured = capsys.readouterr().out
+    with open(path.join(SCRIPT_DIR, 'datacite_ref.xml')) as datacite_ref:
+        # remove the last newline character added by the print function
+        assert captured[:-1] == datacite_ref.read()
