### 2024-12-13 2898498
Add tests for create_release dict
