### 2024-12-13 2737548
Fix remove_doubles when entry is not a list
