### 2024-12-16 3e58a3a
Ensure DATACITE_PATH not defined in env when testing console output
