### 2024-12-16 4aaec40
Add tests for RADAR scripts
