+++ facile_rs/tests/test_create_radar.py
@@ -1,0 +1,104 @@
+import json
+import pytest
+import sys
+from os import path
+
+from facile_rs.create_radar import main
+
+SCRIPT_DIR = path.dirname(path.realpath(__file__))
+METADATA_DIR = path.join(path.dirname(SCRIPT_DIR), 'utils', 'metadata', 'tests')
+
+CODEMETA_LOCATION = path.join(METADATA_DIR, 'codemeta_test.json')
+RADAR_URL = 'https://radar.kit.edu'
+RADAR_USERNAME = 'testuser'
+RADAR_PASSWORD = 'testpassword'
+RADAR_CLIENT_ID = 'testclientid'
+RADAR_CLIENT_SECRET = 'testclientsecret'
+RADAR_WORKSPACE_ID = 'testworkspaceid'
+RADAR_REDIRECT_URL = 'https://example.com'
+RADAR_EMAIL = 'testuser@example.com'
+RADAR_BACKLINK = 'https://example.com'
+ASSETS = [CODEMETA_LOCATION, 'https://www.rfc-editor.org/rfc/rfc2606.txt']
+
+
+def test_error_radar_path_exists(monkeypatch, tmp_path, capsys):
+    radar_path = tmp_path / 'radar_path'
+    radar_path.mkdir()
+    monkeypatch.setattr('sys.argv',
+                        [
+                            sys.argv[0],
+                            '--codemeta-location', CODEMETA_LOCATION,
+                            '--radar-path', str(radar_path),
+                            '--radar-url', RADAR_URL,
+                            '--radar-username', RADAR_USERNAME,
+                            '--radar-password', RADAR_PASSWORD,
+                            '--radar-client-id', RADAR_CLIENT_ID,
+                            '--radar-client-secret', RADAR_CLIENT_SECRET,
+                            '--radar-workspace-id', RADAR_WORKSPACE_ID,
+                            '--radar-redirect-url', RADAR_REDIRECT_URL,
+                            '--radar-email', RADAR_EMAIL,
+                            '--radar-backlink', RADAR_BACKLINK,
+                            '--dry'
+                        ])
+    with pytest.raises(SystemExit, match='^2$'):
+        main()
+    captured = capsys.readouterr().err
+    assert 'already exists' in captured
+
+
+def test_dry_cli(monkeypatch, tmp_path, capsys):
+    """
+    Test that the function runs with CLI options, that CodeMeta metadata is retrieved and that the
+    Radar metadata is printed to stdout in dry mode.
+    """
+    radar_path = tmp_path / 'radar_path'
+    monkeypatch.setattr('sys.argv',
+                        [
+                            sys.argv[0],
+                            '--codemeta-location', CODEMETA_LOCATION,
+                            '--radar-path', str(radar_path),
+                            '--radar-url', RADAR_URL,
+                            '--radar-username', RADAR_USERNAME,
+                            '--radar-password', RADAR_PASSWORD,
+                            '--radar-client-id', RADAR_CLIENT_ID,
+                            '--radar-client-secret', RADAR_CLIENT_SECRET,
+                            '--radar-workspace-id', RADAR_WORKSPACE_ID,
+                            '--radar-redirect-url', RADAR_REDIRECT_URL,
+                            '--radar-email', RADAR_EMAIL,
+                            '--radar-backlink', RADAR_BACKLINK,
+                            '--dry'
+                        ])
+    main()
+    captured = json.loads(capsys.readouterr().out)
[…]
