+++ facile_rs/prepare_radar.py
@@ -20,6 +20,7 @@ Usage
 """
 
 import argparse
+import json
 from pathlib import Path
 
 from .utils import cli, settings
@@ -128,6 +129,8 @@ def main():
             Path(settings.CODEMETA_LOCATION).expanduser().write_text(codemeta.to_json())
         else:
             print(dataset)
+    else:
+        print(json.dumps(radar_dict))
 
 
 def main_deprecated():
