### 2024-12-16 48dd6f1
Fix argv in prepare_release test
