### 2024-12-16 ca54121
Add fixture for ensuring deletion of FACILE-RS env variables before running tests
