+++ facile_rs/tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-@pytest.fixture(autouse=True)
+@pytest.fixture(scope="session", autouse=True)
 def unset_facilers_env(monkeypatch):
     env_vars = [
         'ASSETS',
