### 2024-12-16 245479d
Fix fixture scope
