### 2024-12-16 9604dac
Add tests for prepare_zenodo
