### 2024-12-16 #7330 @exaexa
Hi @MarieHouillon ,

I finally got time to do another review run, and I think this is good to go; accordingly all boxes above are ticked.

Below I've got some very minor nitpicks regarding the paper text flow. These don't need to be fixed/applied but might be considered to improve readability. (I'm not a native speaker so some of these might be completely wrong, feel free to ignore.)

- L8: "FACILE-RS (...) Python package" is split by a veeeeery long parenthesis, to the point that I lost context and expected a verb after the parenthesis. Perhaps split as "Python package FACILE-RS (...) provides ..." (saving 1 word for the determiner)
- L10: ", which" -> " that"
- L12: "makes it easier..." -> "simplifies publishing and ..."
- L15: "creation of releases" does not parse super well to me, "release" in this context has a very specific meaning that is quite far from the common one, maybe "automate the releases of versioned packages" ? (/me unsure)
- L27: "While" sounds like some negative is coming in the next clause... maybe just focus on the main message. My wild shot: "The package integrates well into GitLab CI/CD and similar workflow automation systems, which reduces the need for manual interventions and provides consistent maintainance of metadata and persistent software releases."
- L54: "and an obstacle" clause might be better marked as a consequence using "and thus an obstacle" or so (if I understood that right)
- L58-59: post-conditional "without an automated process" might be better reversed to "However, manual publishing of releases quickly becomes tedious and prone to errors."
- L65: "we aim to provide" -> "we provide" :smile:
- L65: ", which" -> " that"
- L67: I'm not sure if the meaning of "according to" is OK here, might be better go with "following the FAIR4RS ..." ?
- L67: "reputable" might be hard to quantify&interpret, perhaps better go with just "popular" or so
- L69: "such as ... " -> ", including ..." (Not really sure here. To me "such as" somehow implies a cluster of similar items, OTOH here we just list a few examples and their connection and similarity is mostly irrelevant for the reader.)
- L75: "CodeMeta generator" might deserve a printable reference to the concrete one (I imagine googling a "codemeta generator" in 5 years from now will give you a plethora of generators that work, and another plethora of ones that you'd rather not recommend.) -- if there's a citation or a stable link for that (and some space still left for it!) you might want to add one.
- L84: "This" sounds a bit too detached from what it's pointing to, maybe "The" would read better.
- L86: "(e.g. ... for triggering ...)" might read better as "(e.g., ... triggers ...)"

Best & thanks for the software!
-mk

PS Great there's a github port now. You might hear from me in a few months when I'm trying to combine this with JuliaRegistrator release process. :sweat_smile:
