### 2024-12-17 19042ff
Revert "Update CHANGELOG to version 3.1.0"
This reverts commit 4d351a89ab93c5ddae55a988fe3fb3bd9112c7f4
