### 2024-12-18 5b2127d
Update CHANGELOG to version 3.1.0
(cherry picked from commit 4d351a89ab93c5ddae55a988fe3fb3bd9112c7f4)

Co-authored-by: Marie Houillon <marie.houillon@kit.edu>
