### 2024-12-18 291cf6f
Update FACILE-RS and add reference to templates in paper
