+++ pyproject.toml
@@ -30,6 +30,7 @@ classifiers = [
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
     'Programming Language :: Python :: 3.12',
+    'Programming Language :: Python :: 3.13',
 ]
 dependencies = [
     "bagit~=1.7",
@@ -39,7 +40,8 @@ dependencies = [
     "PyYaml~=5.3.1",
     "requests~=2.23",
     "python-resize-image~=1.1.20",
-    "Pillow~=9.4.0",
+    "Pillow~=9.4.0; python_version <= '3.11'",
+    "Pillow~=11.0; python_version > '3.11'",
     "setuptools",
 ]
 dynamic = ["version"]
