### 2024-12-18 #7330 @MarieHouillon
Hi @exaexa, thank you very much for your review!
We have updated FACILE-RS to the latest version in the joss branch, included some of your suggestions in the paper and also added a mention of the templates.
