### 2025-01-06 7caa3eb
Assets token fix
