+++ facile_rs/utils/http.py
@@ -19,10 +19,11 @@ def fetch_files(locations, path, headers={}):
     :type path: str
     """
     for location in locations:
-        target = path / location.split('/')[-1]
+        parsed_url = urlparse(location)
+        target = path / parsed_url.path.split('/')[-1]
 
         logger.debug('location = %s, target = %s', location, target)
-        if urlparse(location).scheme:
+        if parsed_url.scheme:
             response = requests.get(location, headers=headers)
             response.raise_for_status()
             with open(target, 'wb') as f:
