### 2025-01-07 21bc466
Fix when author or contributor is not a list
