### 2025-01-07 38f7371
Fix wrong location in CHANGELOG.md
