### 2025-01-07 a77850b
Update CHANGELOG.md for v3.2.0
