+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## Unreleased
+## v3.2.0
 
 ### Added
 - Add --assets-token and --assets-token-name arguments for fetching assets from private repositories
