+++ CITATION.cff
@@ -5,9 +5,9 @@ type: software
 title: FACILE-RS
 abstract: A set of Python scripts for tasks around the archival and long term preservation
   of software repositories.
-doi: 10.35097/ay82j7v07m1zvhwv
-version: v3.1.0
-date-released: 2024-12-18
+doi: 10.35097/ta2t9z8ze6m9mhts
+version: v3.2.0
+date-released: 2025-01-07
 authors:
 - name: openCARP consortium
 - given-names: Marie
@@ -56,6 +56,6 @@ preferred-citation:
     given-names: Robert
     orcid: https://orcid.org/0000-0001-9063-2703
 identifiers:
-- description: This is the archived snapshot of version v3.1.0 of FACILE-RS
+- description: This is the archived snapshot of version v3.2.0 of FACILE-RS
   type: doi
-  value: 10.35097/ay82j7v07m1zvhwv
+  value: 10.35097/ta2t9z8ze6m9mhts
