### 2025-01-15 8672acb
Make command-line option match environement variable name
