+++ facile_rs/create_bagpack.py
@@ -36,7 +36,7 @@ def create_parser(add_help=True):
                         help='Assets to be added to the bag.')
     parser.add_argument('--bag-path', dest='bag_path',
                         help='Path to the Bag directory')
-    parser.add_argument('--bag-info-location', dest='bag_info_locations', action='append', default=[],
+    parser.add_argument('--bag-info-locations', '--bag-info-location', dest='bag_info_locations', action='append', default=[],
                         help='Locations of the bag-info YAML/JSON files')
     parser.add_argument('--datacite-path', dest='datacite_path',
                         help='Path to the DataCite XML file')
