### 2025-01-20 #7330 @atrisovic
Hello @RobLBaker, @AngryMaciek,

I hope all is well. Just a gentle reminder about FACILE-RS: archival and long-term preservation of research software repositories made easy. It would be great to get an update on your progress with the revisions. One review has been already completed, but we are not able to complete this without your input! Please let us know if you need any clarification or additional support. Thanks for your efforts!
Ana
