### 2025-01-21 fd555ce
Add run_notebook_pipeline
