### 2025-01-21 #7330 @RobLBaker
I will try to get to it this week.

Best,

Rob

Robert Baker, Ph.D. (he/him)
Data Scientist/Integrative Biologist
National Park Service Inventory & Monitoring Division
1201 Oakridge Drive, Suite 150
Fort Collins, CO 80525
***@***.*** https://orcid.org/0000-0001-7591-5035

From: Ana Trisovic ***@***.***>
Sent: Monday, January 20, 2025 11:57 AM
Cc: Baker, Robert L ***@***.***>; Mention ***@***.***>
Subject: [EXTERNAL] Re: [openjournals/joss-reviews] [REVIEW]: FACILE-RS: archival and long term preservation of research software repositories made easy (Issue #7330)

 This email has been received from outside of DOI - Use caution before clicking on links, opening attachments, or responding.

Hello @RobLBaker<https://github.com/RobLBaker>, @AngryMaciek<https://github.com/AngryMaciek>,

I hope all is well. Just a gentle reminder about FACILE-RS: archival and long-term preservation of research software repositories made easy. It would be great to get an update on your progress with the revisions. One review has been already completed, but we are not able to complete this without your input! Please let us know if you need any clarification or additional support. Thanks for your efforts!
Ana

-
