### 2025-01-27 a685bb0
Remove deprecation handling from run_notebook_pipeline
