### 2025-02-18 #7330 @RobLBaker
I could be doing something wrong, but when I run `pip install FACILE-RS` as indicated in the readme, I get the following errors:

Defaulting to user installation because normal site-packages is not writeable
Collecting FACILE-RS
  Using cached FACILE_RS-3.2.0-py3-none-any.whl.metadata (25 kB)
Collecting bagit~=1.7 (from FACILE-RS)
  Using cached bagit-1.8.1-py2.py3-none-any.whl.metadata (8.3 kB)
Collecting pypandoc~=1.6 (from FACILE-RS)
  Using cached pypandoc-1.15-py3-none-any.whl.metadata (16 kB)
Collecting python-dotenv~=0.10 (from FACILE-RS)
  Using cached python_dotenv-0.21.1-py3-none-any.whl.metadata (21 kB)
Collecting python-frontmatter~=0.4 (from FACILE-RS)
  Using cached python_frontmatter-0.5.0-py3-none-any.whl.metadata (3.0 kB)
Collecting PyYaml~=5.3.1 (from FACILE-RS)
  Using cached PyYAML-5.3.1.tar.gz (269 kB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [34 lines of output]
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/tmp/pip-install-wiat06gu/pyyaml_cfba5b58e7f340978e9e49749768aeac/setup.py", line 291, in <module>
          setup(
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/core.py", line 186, in setup
          return run_commands(dist)
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/core.py", line 202, in run_commands
          dist.run_commands()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/dist.py", line 983, in run_commands
          self.run_command(cmd)
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/dist.py", line 999, in run_command
          super().run_command(command)
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/dist.py", line 1002, in run_command
          cmd_obj.run()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/command/egg_info.py", line 312, in run
          self.find_sources()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/command/egg_info.py", line 320, in find_sources
          mm.run()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/command/egg_info.py", line 543, in run
          self.add_defaults()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/command/egg_info.py", line 581, in add_defaults
          sdist.add_defaults(self)
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/command/sdist.py", line 109, in add_defaults
          super().add_defaults()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/command/sdist.py", line 239, in add_defaults
          self._add_defaults_ext()
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/command/sdist.py", line 324, in _add_defaults_ext
          self.filelist.extend(build_ext.get_source_files())
        File "/tmp/pip-install-wiat06gu/pyyaml_cfba5b58e7f340978e9e49749768aeac/setup.py", line 199, in get_source_files
          self.cython_sources(ext.sources, ext)
        File "/opt/python/3.9.12/lib/python3.9/site-packages/setuptools/_distutils/cmd.py", line 120, in __getattr__
          raise AttributeError(attr)
      AttributeError: cython_sources
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
Note: you may need to restart the kernel to use updated packages.
