### 2025-02-18 #7330 @exaexa
@RobLBaker I managed to install with python 3.13 (see log below). Maybe there's proper version bounds missing somewhere in the package metadata. Good catch, this might require attention.

Log from my machine (debian testing/trixie, fully updated sometime last week):

```
(facile) ~ $ pip install FACILE-RS
Collecting FACILE-RS
  Downloading FACILE_RS-3.2.0-py3-none-any.whl.metadata (25 kB)
Collecting bagit~=1.7 (from FACILE-RS)
  Downloading bagit-1.8.1-py2.py3-none-any.whl.metadata (8.3 kB)
Collecting pypandoc~=1.6 (from FACILE-RS)
  Downloading pypandoc-1.15-py3-none-any.whl.metadata (16 kB)
Collecting python-dotenv~=0.10 (from FACILE-RS)
  Downloading python_dotenv-0.21.1-py3-none-any.whl.metadata (21 kB)
Collecting python-frontmatter~=0.4 (from FACILE-RS)
  Downloading python_frontmatter-0.5.0-py3-none-any.whl.metadata (3.0 kB)
Collecting PyYaml~=5.3.1 (from FACILE-RS)
  Downloading PyYAML-5.3.1.tar.gz (269 kB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting requests~=2.23 (from FACILE-RS)
  Using cached requests-2.32.3-py3-none-any.whl.metadata (4.6 kB)
Collecting python-resize-image~=1.1.20 (from FACILE-RS)
  Downloading python_resize_image-1.1.20-py2.py3-none-any.whl.metadata (9.0 kB)
Collecting Pillow~=11.0 (from FACILE-RS)
  Using cached pillow-11.1.0-cp313-cp313-manylinux_2_28_x86_64.whl.metadata (9.1 kB)
Collecting setuptools (from FACILE-RS)
  Using cached setuptools-75.8.0-py3-none-any.whl.metadata (6.7 kB)
Collecting six (from python-frontmatter~=0.4->FACILE-RS)
  Using cached six-1.17.0-py2.py3-none-any.whl.metadata (1.7 kB)
Collecting charset-normalizer<4,>=2 (from requests~=2.23->FACILE-RS)
  Using cached charset_normalizer-3.4.1-cp313-cp313-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (35 kB)
Collecting idna<4,>=2.5 (from requests~=2.23->FACILE-RS)
  Using cached idna-3.10-py3-none-any.whl.metadata (10 kB)
Collecting urllib3<3,>=1.21.1 (from requests~=2.23->FACILE-RS)
  Using cached urllib3-2.3.0-py3-none-any.whl.metadata (6.5 kB)
Collecting certifi>=2017.4.17 (from requests~=2.23->FACILE-RS)
  Using cached certifi-2025.1.31-py3-none-any.whl.metadata (2.5 kB)
Downloading FACILE_RS-3.2.0-py3-none-any.whl (71 kB)
Downloading bagit-1.8.1-py2.py3-none-any.whl (35 kB)
Using cached pillow-11.1.0-cp313-cp313-manylinux_2_28_x86_64.whl (4.5 MB)
Downloading pypandoc-1.15-py3-none-any.whl (21 kB)
Downloading python_dotenv-0.21.1-py3-none-any.whl (19 kB)
Downloading python_frontmatter-0.5.0-py3-none-any.whl (8.4 kB)
Downloading python_resize_image-1.1.20-py2.py3-none-any.whl (8.4 kB)
Using cached requests-2.32.3-py3-none-any.whl (64 kB)
Using cached setuptools-75.8.0-py3-none-any.whl (1.2 MB)
Using cached certifi-2025.1.31-py3-none-any.whl (166 kB)
Using cached charset_normalizer-3.4.1-cp313-cp313-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (144 kB)
Using cached idna-3.10-py3-none-any.whl (70 kB)
Using cached urllib3-2.3.0-py3-none-any.whl (128 kB)
Using cached six-1.17.0-py2.py3-none-any.whl (11 kB)
Building wheels for collected packages: PyYaml
  Building wheel for PyYaml (pyproject.toml) ... done
  Created wheel for PyYaml: filename=PyYAML-5.3.1-cp313-cp313-linux_x86_64.whl size=44694 sha256=50d66acdd82b2c155869e4ef8d25501d632d33f4d30cc63a1628bc0dc9fe0355
  Stored in directory: /home/exa/.cache/pip/wheels/1e/db/36/42f3881eef0b64b33b8382d2d1638b90e688a8e3fffd6180c9
Successfully built PyYaml
Installing collected packages: bagit, urllib3, six, setuptools, PyYaml, python-dotenv, pypandoc, Pillow, idna, charset-normalizer, certifi, requests, python-frontmatter, python-resize-image, FACILE-RS
Successfully installed FACILE-RS-3.2.0 Pillow-11.1.0 PyYaml-5.3.1 bagit-1.8.1 certifi-2025.1.31 charset-normalizer-3.4.1 idna-3.10 pypandoc-1.15 python-dotenv-0.21.1 python-frontmatter-0.5.0 python-resize-image-1.1.20 requests-2.32.3 setuptools-75.8.0 six-1.17.0 urllib3-2.3.0

(facile) ~ $ pip --version
pip 25.0 from /tmp/facile/lib/python3.13/site-packages/pip (python 3.13)

(facile) ~ $ python --version
Python 3.13.1
```
