### 2025-02-21 74dcb4d
Add details to docstring for run_notebook_pipeline.py
