### 2025-03-01 6dd989a
Add catalog pipeline
