### 2025-04-01 53c4fd4
Add send_mail utils function
