### 2025-04-01 6deb912
Refactor create_bagpack and rename DATACITE_PATH -> DATACITE_LOCATIONS
