+++ pyproject.toml
@@ -35,8 +35,8 @@ classifiers = [
 dependencies = [
     "bagit~=1.7",
     "pypandoc~=1.6",
-    "python-dotenv~=0.10",
-    "python-frontmatter~=0.4",
+    "python-dotenv~=1.1",
+    "python-frontmatter~=1.1",
     "PyYaml~=6.0.2",
     "requests~=2.23",
     "python-resize-image~=1.1.20",
