### 2025-04-01 a7698f6
Add custom Parser and refactor create_bag
