+++ facile_rs/utils/cli.py
@@ -1,7 +1,11 @@
 import argparse
-import inspect
-import os.path
+import logging
+import os
+import sys
+from pathlib import Path
 from warnings import warn
+
+from dotenv import load_dotenv
 
 from facile_rs import (
     create_bag,
@@ -20,6 +24,39 @@ from facile_rs import (
 )
 
 
+class Parser(argparse.ArgumentParser):
+    """
+    Custom ArgumentParser class, which adds the functionality to add default values from the environment.
+    """
+
+    def add_argument(self, *args, **kwargs):
+        argument = super().add_argument(*args, **kwargs)
+
+        if not isinstance(argument, argparse._HelpAction):
+            # update the default value for the argument with a possible environment variable
+            default = os.getenv(argument.dest)
+            if default is not None:
+                # if the argument is a list, split the default from the environment
+                if kwargs.get('nargs') == '*' or kwargs.get('action') == 'append':
+                    default = default.split()
+
+                argument.default = default
+
+            # remove the required argument if a default is set
+            if argument.required and argument.default:
+                argument.required = False
+
+        return argument
+
+    def get_subparser(self, subcommand):
+        # returns the subparser for a given subcommand
+        for action in self._actions:
+            if isinstance(action, argparse._SubParsersAction):
+                subparser = action.choices[subcommand]
+                return subparser
+
+
+
 # create the top-level parser
 def create_parser():
     """
@@ -29,8 +66,9 @@ def create_parser():
 
     :return: The parser object.
     """
+
     # Main parser
-    parser = argparse.ArgumentParser(prog='facile-rs',
+    parser = Parser(prog='facile-rs',
         description="FACILE-RS command-line tool, to perform metadata conversion and software publication" \
             "based on CodeMeta metadata.",
         epilog="Get help on a subcommand by running 'facile-rs <subcommand> -h'.")
@@ -46,7 +84,7 @@ def create_parser():
                                                            help='Update CodeMeta file with the given version and date',
                                                            parents=[prepare_release.create_parser(add_help=False)],
                                                            add_help=True)
-    parser_release_prepare.set_defaults(func=prepare_release.main)
+    parser_release_prepare.set_defaults(module=prepare_release)
 
     # Parser for the 'gitlab' subcommand
     parser_gitlab = subparsers.add_parser('gitlab', help='Perform operations for GitLab releases')
@@ -55,7 +93,7 @@ def create_parser():
                                                         parents=[create_release.create_parser(add_help=False)],
                                                         help="Create a release on GitLab",
                                                         add_help=True)
-    parser_gitlab_publish.set_defaults(func=create_release.main)
+    parser_gitlab_publish.set_defaults(module=create_release)
 
     # Parser for the 'radar' subcommand
     parser_radar = subparsers.add_parser('radar', help='Perform operations for RADAR releases')
@@ -64,12 +102,12 @@ def create_parser():
                                                        help='Prepare a release on RADAR',
                                                        parents=[prepare_radar.create_parser(add_help=False)],
                                                        add_help=True)
-    parser_radar_prepare.set_defaults(func=prepare_radar.main)
+    parser_radar_prepare.set_defaults(module=prepare_radar)
     parser_radar_upload = radar_subparsers.add_parser('upload',
                                                       help='Create a release on RADAR',
                                                       parents=[create_radar.create_parser(add_help=False)],
                                                       add_help=True)
-    parser_radar_upload.set_defaults(func=create_radar.main)
+    parser_radar_upload.set_defaults(module=create_radar)
 
     # Parser for the 'zenodo' subcommand
[…]
