+++ facile_rs/main.py
@@ -26,37 +26,34 @@ To generate CFF metadata from a CodeMeta metadata file:
     $ facile-rs cff create --codemeta-location codemeta.json
 """
 
-import argparse
-import sys
-
 from .utils import cli
 
 
 def main():
+    # load the environment
+    cli.setup_env()
+
     # Create the command-line parser
     parser = cli.create_parser()
     args = parser.parse_args()
+
+    # setup logs
+    cli.setup_logs(args.LOG_LEVEL, args.LOG_FILE)
+
     # Print help if subcommand is missing
     subcommand = args.subcommand
     if subcommand is None:
         parser.print_help()
-        sys.exit(1)
-    # args.mode contains the name of the script to execute
-    try:
-        func = args.func
-    except AttributeError:
-        print(f"Error: missing subcommand for '{subcommand}'.")
-        # Get help for given subcommand
-        for action in parser._actions:
-            if isinstance(action, argparse._SubParsersAction):
-                subparser = action.choices[subcommand]
-                subparser.print_help()
-                break
-        sys.exit(1)
-    # Call the script with the remaining arguments
-    command_str = ' '.join(sys.argv[:3])
-    sys.argv = [command_str] + sys.argv[3:]
-    func()
+    else:
+        try:
+            module = args.module
+        except AttributeError:
+            subparser = parser.get_subparser(subcommand)
+            subparser.print_help()
+        else:
+            # Call the "main" function of the module
+            module.main(args)
+
 
 if __name__ == "__main__":
     main()
