+++ .gitlab/ci/pytest.gitlab-ci.yml
@@ -7,4 +7,4 @@ test-install:
     - pytest -v
   parallel:
     matrix:
-      - PYTHON_VERSION: ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
+      - PYTHON_VERSION: ["3.9", "3.10", "3.11", "3.12", "3.13"]
