### 2025-04-04 b847b6c
Fix create_bagpack and add fetch_file method
