### 2025-04-04 a0e0879
Rename mkenv -> mk_empty_dir
