+++ facile_rs/create_bagpack.py
@@ -24,7 +24,7 @@ from pathlib import Path
 
 import bagit
 
-from .utils import cli, mkdir, settings
+from .utils import cli, mk_empty_dir, settings
 from .utils.checksum import get_sha256, get_sha512
 from .utils.http import fetch_dict, fetch_files
 
@@ -65,7 +65,7 @@ def main():
     # setup the bag
     bag_path = Path(settings.BAG_PATH).expanduser()
     try:
-        mkdir(bag_path, settings.OVERWRITE)
+        mk_empty_dir(bag_path, settings.OVERWRITE)
     except FileExistsError:
         parser.error(f'{bag_path} already exists.')
 
