### 2025-04-04 af8152f
Refactor mk_empty_dir -> setup_assets_path and add setup_tmp_assets_path
