### 2025-04-04 fa12503
Improve utils.mail
