+++ facile_rs/create_radar.py
@@ -43,7 +43,7 @@ def create_parser(add_help=True):
                         help='Do not sort authors alphabetically, keep order in codemeta.json file')
     parser.set_defaults(sort_authors=True)
     parser.add_argument('--radar-path', dest='radar_path',
-                        help='Path to the local directory, where the assets are collected before upload.')
+                        help='Path to the local directory, where the assets are collected before upload. Optional: if not provided, a temporary directory is used.')
     parser.add_argument('--radar-url', dest='radar_url',
                         help='URL of the RADAR service.')
     parser.add_argument('--radar-username', dest='radar_username',
