+++ facile_rs/create_zenodo.py
@@ -43,7 +43,7 @@ def create_parser(add_help=True):
                         help='Do not sort authors alphabetically, keep order in codemeta.json file')
     parser.set_defaults(sort_authors=True)
     parser.add_argument('--zenodo-path', dest='zenodo_path',
-                        help='Path to the local directory, where the assets are collected before upload.')
+                        help='Path to the local directory, where the assets are collected before upload. Optional: if not provided, a temporary directory is used.')
     parser.add_argument('--zenodo-url', dest='zenodo_url',
                         help='URL of the Zenodo service. Test environment available at https://sandbox.zenodo.org')
     parser.add_argument('--zenodo-token', dest='zenodo_token',
