### 2025-04-10 516f464
Allow for existing directories for create_radar and create_zenodo, refactor setup_assets_path, and add AssetExistsError
