### 2025-04-17 a19d81a
Add changes proposed to Jochen
