### 2025-05-01 60b75ef
Fix setup_logs
