### 2025-05-01 673f67a
Move main to utils.cli
