+++ facile_rs/utils/cli.py
@@ -1,3 +1,29 @@
+"""FACILE-RS command-line tool, used to call the different scripts of the FACILE-RS project.
+
+Description
+-----------
+
+This script is the entry point of the FACILE-RS command-line tool.
+It is used to call the different scripts of the FACILE-RS project.
+Use subcommands to select a platform (Zenodo, RADAR, ...) or metadata type (CFF, DataCite,...).
+Use subsubcommands to select a FACILE-RS functionality.
+
+Usage
+-----
+
+.. argparse::
+    :module: facile_rs.utils.cli
+    :func: create_parser
+    :prog: facile-rs
+
+Examples
+--------
+
+To generate CFF metadata from a CodeMeta metadata file:
+
+    $ facile-rs cff create --codemeta-location codemeta.json
+"""
+
 import argparse
 import logging
 import os
@@ -54,7 +80,6 @@ class Parser(argparse.ArgumentParser):
             if isinstance(action, argparse._SubParsersAction):
                 subparser = action.choices[subcommand]
                 return subparser
-
 
 
 # create the top-level parser
@@ -183,6 +208,42 @@ def create_parser():
     return parser
 
 
+def setup_env():
+    load_dotenv(Path().cwd() / '.env')
+
+
+def setup_logs(log_level, log_file):
+    log_level = log_level.upper()
+    log_file = Path(log_file).expanduser().as_posix() if log_file is not None else None
+    logging.basicConfig(level=log_level, filename=log_file, format='[%(asctime)s] %(levelname)s %(name)s: %(message)s')
+
+
+def main():
+    # load the environment
+    setup_env()
+
+    # Create the command-line parser
+    parser = create_parser()
+    args = parser.parse_args()
+
+    # Print help if subcommand is missing
+    subcommand = args.subcommand
+    if subcommand is None:
+        parser.print_help()
+    else:
+        try:
+            module = args.module
+        except AttributeError:
+            subparser = parser.get_subparser(subcommand)
+            subparser.print_help()
+        else:
+            # setup logs
+            setup_logs(args.LOG_LEVEL, args.LOG_FILE)
+
+            # Call the "main" function of the module
+            module.main(args)
+
+
 def cli_call_deprecated(module_name):
     """
     Display a deprecation warning when a script is called from the command line directly,
@@ -208,13 +269,3 @@ def cli_call_deprecated(module_name):
 
     # call the main function
     module.main(args)
-
-
-def setup_env():
-    load_dotenv(Path().cwd() / '.env')
-
-
-def setup_logs(log_level, log_file):
-    log_level = log_level.upper()
-    log_file = Path(log_file).expanduser().as_posix() if log_file is not None else None
-    logging.basicConfig(level=log_level, filename=log_file, format='[%(asctime)s] %(levelname)s %(name)s: %(message)s')
