### 2025-05-01 91dcad0
Add ParserError exception
