### 2025-05-01 dee3cfc
Add description to bag subparsers
