### 2025-05-06 43e8a46
Add more complete test suite and coverage
