### 2025-05-06 39534cd
facile_rs.main -> facile_rs.utils.cli
