+++ facile_rs/tests/test_create_bag.py
@@ -2,7 +2,7 @@ import pytest
 import sys
 from os import path
 
-from facile_rs.main import main
+from facile_rs.utils.cli import main
 from facile_rs.utils.http import fetch_dict
 
 
