### 2025-05-07 #7330 @danielskatz
@atrisovic - again, can you help me understand what's happening here?
