### 2025-05-09 9bff5f5
Refactor all scripts
