+++ facile_rs/create_radar.py
@@ -19,12 +19,10 @@ Usage
     :prog: create_radar.py
 
 """
-
-import argparse
 import json
 
-from .utils import cli, settings, setup_assets_path, setup_tmp_assets_path
-from .utils.exceptions import AssetExistsError
+from .utils import cli, setup_assets_path, setup_tmp_assets_path
+from .utils.exceptions import AssetExistsError, ParserError
 from .utils.http import fetch_files
 from .utils.mail import send_mail
 from .utils.metadata import CodemetaMetadata, RadarMetadata
@@ -32,96 +30,84 @@ from .utils.radar import create_radar_dataset, fetch_radar_token, update_radar_d
 
 
 def create_parser(add_help=True):
-    parser = argparse.ArgumentParser(add_help=add_help)
-    parser.add_argument('assets', nargs='*', default=[],
+    parser = cli.Parser(add_help=add_help)
+
+    parser.add_argument('ASSETS', nargs='*', default=[],
                         help='Assets to be added to the repository.')
-    parser.add_argument('--codemeta-location', dest='codemeta_location',
+    parser.add_argument('--codemeta-location', dest='CODEMETA_LOCATION', required=True,
                         help='Location of the main codemeta.json JSON file')
-    parser.add_argument('--creators-locations', '--creators-location', dest='creators_locations', action='append', default=[],
+    parser.add_argument('--creators-locations', '--creators-location', dest='CREATORS_LOCATIONS',
+                        action='append', default=[],
                         help='Locations of codemeta JSON files for additional creators')
-    parser.add_argument('--contributors-locations', '--contributors-location', dest='contributors_locations', action='append', default=[],
+    parser.add_argument('--contributors-locations', '--contributors-location', dest='CONTRIBUTORS_LOCATIONS',
+                        action='append', default=[],
                         help='Locations of codemeta JSON files for additional contributors')
-    parser.add_argument('--no-sort-authors', dest='sort_authors', action='store_false',
+    parser.add_argument('--no-sort-authors', dest='SORT_AUTHORS', action='store_false', default=True,
                         help='Do not sort authors alphabetically, keep order in codemeta.json file')
-    parser.set_defaults(sort_authors=True)
-    parser.add_argument('--radar-path', dest='radar_path',
-                        help='Path to the local directory, where the assets are collected before upload. Optional: if not provided, a temporary directory is used.')
-    parser.add_argument('--radar-url', dest='radar_url',
+    parser.add_argument('--radar-path', dest='RADAR_PATH',
+                        help='Path to the local directory, where the assets are collected before upload. '
+                             'Optional: if not provided, a temporary directory is used.')
+    parser.add_argument('--radar-url', dest='RADAR_URL', required=True,
                         help='URL of the RADAR service.')
-    parser.add_argument('--radar-username', dest='radar_username',
+    parser.add_argument('--radar-username', dest='RADAR_USERNAME', required=True,
                         help='Username for the RADAR service.')
-    parser.add_argument('--radar-password', dest='radar_password',
+    parser.add_argument('--radar-password', dest='RADAR_PASSWORD', required=True,
                         help='Password for the RADAR service.')
[…]
