+++ facile_rs/run_bibtex_pipeline.py
@@ -18,15 +18,13 @@ Usage
     :prog: run_bibtex_pipeline.py
 
 """
-
-import argparse
 import logging
 from pathlib import Path
 
 import frontmatter
 import pypandoc
 
-from .utils import cli, settings
+from .utils import cli
 from .utils.grav import collect_pages
 
 logger = logging.getLogger(__file__)
@@ -39,40 +37,32 @@ nocite: '@*'
 
 
 def create_parser(add_help=True):
-    parser = argparse.ArgumentParser(add_help=add_help)
+    parser = cli.Parser(add_help=add_help)
 
-    parser.add_argument('--grav-path', dest='grav_path',
+    parser.add_argument('--grav-path', dest='GRAV_PATH', required=True,
                         help='Path to the grav repository directory.')
-    parser.add_argument('--pipeline', dest='pipeline',
+    parser.add_argument('--pipeline', dest='PIPELINE', required=True,
                         help='Name of the pipeline as specified in the GRAV metadata.')
-    parser.add_argument('--pipeline-source', dest='pipeline_source',
+    parser.add_argument('--pipeline-source', dest='PIPELINE_SOURCE', required=True,
                         help='Path to the source directory for the pipeline.')
-    parser.add_argument('--pipeline-csl', dest='pipeline_csl',
+    parser.add_argument('--pipeline-csl', dest='PIPELINE_CSL',
                         help='Path to the source directory for the pipeline.')
-    parser.add_argument('--log-level', dest='log_level',
+    parser.add_argument('--log-level', dest='LOG_LEVEL', default='WARN',
                         help='Log level (ERROR, WARN, INFO, or DEBUG)')
-    parser.add_argument('--log-file', dest='log_file',
+    parser.add_argument('--log-file', dest='LOG_FILE',
                         help='Path to the log file')
     return parser
 
 
-def main():
-    parser = create_parser()
-
-    settings.setup(parser, validate=[
-        'GRAV_PATH',
-        'PIPELINE',
-        'PIPELINE_SOURCE'
-    ])
-
+def main(args):
     # loop over the found pages and write the content into the files
-    for page_path, page, source in collect_pages(settings.GRAV_PATH, settings.PIPELINE):
-        source_path = Path(settings.PIPELINE_SOURCE).expanduser() / source
+    for page_path, page, source in collect_pages(args.GRAV_PATH, args.PIPELINE):
+        source_path = Path(args.PIPELINE_SOURCE).expanduser() / source
         logger.debug('page_path = %s, source_path = %s', page_path, source_path)
 
         extra_args = [f'--bibliography={source_path}', '--citeproc', '--wrap=preserve']
-        if settings.PIPELINE_CSL:
-            extra_args.append(f'--csl={settings.PIPELINE_CSL}')
+        if args.PIPELINE_CSL:
+            extra_args.append(f'--csl={args.PIPELINE_CSL}')
 
         page.content = pypandoc.convert_text(TEMPLATE, to='html', format='md',
                                              extra_args=extra_args)
@@ -82,7 +72,7 @@ def main():
 
 
 def main_deprecated():
-    cli.cli_call_deprecated(main)
+    cli.cli_call_deprecated(__name__)
 
 
 if __name__ == "__main__":

[…]
