### 2025-05-09 6be59a2
Rename cli_call_deprecated -> main_deprecated
