### 2025-05-09 5ff5a7e
Cleanup code and make it ruff compliant again
