+++ facile_rs/tests/test_create_radar.py
@@ -1,7 +1,8 @@
 import json
-import pytest
 import sys
 from os import path
+
+import pytest
 
 from facile_rs.utils.cli import main
 
