### 2025-05-09 0dad476
Typo in module description
