### 2025-05-27 97a046c
HTML mathjax rendering
