### 2025-05-27 6831cdc
Support Zenodo versioning
