+++ facile_rs/prepare_zenodo.py
@@ -9,6 +9,9 @@ This script creates an empty archive in Zenodo in order to reserve a DOI and a Z
 Both are stored in the CodeMeta metadata file provided as input and can be later used by the script ``create_zenodo.py``
 to populate the Zenodo archive.
 
+Optionally, the script can create a new version from an existing Zenodo record.
+See the option --zenodo-version-update for more details.
+
 Usage
 -----
 
@@ -35,6 +38,12 @@ def create_parser(add_help=True):
                         help='URL of the Zenodo service. Test environment available at https://sandbox.zenodo.org')
     parser.add_argument('--zenodo-token', dest='ZENODO_TOKEN', required=True,
                         help='Zenodo personal token.')
+    parser.add_argument('--zenodo-version-update', dest='ZENODO_VERSION_UPDATE', default=None,
+                        help='Enable Zenodo version update. Can be "codemeta" or a Zenodo identifier. '
+                        'If omitted, a new Zenodo dataset is created without versioning. '
+                        'If set to "codemeta", a Zenodo identifier is searched in the CodeMeta file, and a new version is '
+                        'created from it (if found). '
+                        'Any other value is considered as a Zenodo identifier: a new version will be created from it. ')
     parser.add_argument('--dry', action='store_true', dest='DRY',
                         help='Perform a dry run, do not upload anything.')
     parser.add_argument('--log-level', dest='LOG_LEVEL', default='WARN',
@@ -56,9 +65,23 @@ def main(args):
     zenodo_metadata = ZenodoMetadata({'name': name})
     zenodo_dict = zenodo_metadata.as_dict()
 
+    # Management of Zenodo versioning
+    old_zenodo_id = None
+    if args.CODEMETA_LOCATION and args.ZENODO_VERSION_UPDATE == 'codemeta':
+        # Check if CodeMeta file contains a Zenodo identifier
+        identifiers = codemeta.data.get('identifier', [])
+        if not isinstance(identifiers, list):
+            identifiers = [identifiers]
+        for identifier in identifiers:
+            if isinstance(identifier, dict) and identifier.get('propertyID') == 'Zenodo':
+                old_zenodo_id = identifier.get('value', None)
+    elif args.ZENODO_VERSION_UPDATE is not None:
+        # Use the provided Zenodo identifier to create a new version
+        old_zenodo_id = args.ZENODO_VERSION_UPDATE
+
     if not args.DRY:
         # create Zenodo dataset
-        dataset_id = create_zenodo_dataset(args.ZENODO_URL, args.ZENODO_TOKEN, zenodo_dict)
+        dataset_id = create_zenodo_dataset(args.ZENODO_URL, args.ZENODO_TOKEN, zenodo_dict, previous_version=old_zenodo_id)
         dataset = prepare_zenodo_dataset(args.ZENODO_URL, dataset_id, args.ZENODO_TOKEN)
 
         doi = dataset.get('metadata', {}).get('doi', {})
