+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Rename `--datacite-path` to `--datacite-location` in `facile-rs bagpack create` and allow for remote locations.
 - The "publisher" field in Zenodo metadata is now populated with the "publisher" name provided in the Codemeta file.
 - Include MathJax to HTML outputs in `run_docstring_pipeline`, and add option `--mathjax-location` to customize MathJax location or disable it.
+- `facile-rs zenodo prepare` now allows to create a new version of a Zenodo record, if a Zenodo ID is present in the CodeMeta file. This behavior can set using the option `--zenodo-version-update`.
 
 ### Changed
 - The option `--overwrite` is not read from environment anymore: it shall be passed from the command line.
