### 2025-06-06 #7330 @NicolettaNinkovic1
Hi, it is my pleasure to help with this review. I installed the package and reviewed the documentation.

I have a suggestion. It would be great if this would work as well. I tested it to make cff file.
```facile-rs cff create --codemeta-location .
/home/nicoletta/joss/env/bin/bagit.py:24: UserWarning: pkg_resources is deprecated as an API. See https://setuptools.pypa.io/en/latest/pkg_resources.html. The pkg_resources package is slated for removal as early as 2025-11-30. Refrain from using this package or pin to Setuptools<81.
  from pkg_resources import DistributionNotFound, get_distribution
Traceback (most recent call last):
  File "/home/nicoletta/joss/env/bin/facile-rs", line 8, in <module>
    sys.exit(main())
```
this worked `facile-rs cff create --codemeta-location codemeta.json`

I don't see **community guidelines**. Could you please point me in the right direction?
