### 2025-06-06 #7330 @NicolettaNinkovic1
Hello, I read your paper, which is written well and your software is very useful. It is great to see you already have users. In Figure 1, it was not clear to me what you meant by "test: check-readme". Is that a command? If so, I did not see how to run it.
