### 2025-06-06 #7330 @MarieHouillon
Hi @NicolettaNinkovic1 ,

Thank you very much for review. As FACILE-RS has initially been a part of the openCARP project, the community guidelines of [openCARP](https://opencarp.org/community/contribute) apply.
I will add a reference to it in the repository, thank you for raising this issue.

This `check-readme` job is only here to illustrate other test jobs that could be present within a CI workflow. It doesn't correspond to anything within FACILE-RS. I can add a note about this in the paper if this is confusing.
