### 2025-06-24 #7330 @MarieHouillon
Hello,

I've created an archive on RADAR that should comply with the requirements:
- DOI: [10.35097/qpnn3svjjg7w4c5n](http://doi.org/10.35097/qpnn3svjjg7w4c5n)
- Version: 3.3.0
