### 2025-06-25 #7330 @MarieHouillon
> Ok excellent! Just to confirm - everything else from the additional tasks for the author looks good to you?

Yes, it looks good.
