### 2025-06-25 #7330 @atrisovic
Hi @MarieHouillon I wonder if you could incorporate the following changes in the paper:

- FAIR principles defined two times; remove one from the Statement of need
- remove historically
- A commen on if FACILE-RS  supports metadata validation for codemeta, cff and datacite?
- has now been acknowledged → is now widely acknowledged
- long term → long-term
- to follow → follow
