### 2025-06-26 #7330 @MarieHouillon
Hi @atrisovic ,

Thank you for your review. I've updated the paper as suggested. If it looks good to you, I will generate a new archive.
