### 2025-07-28 #7330 @atrisovic
I haven't seen it before, but maybe @danielskatz could help here.
