### 2025-08-01 57adb2b
order commands alphabetically in README
