### 2025-08-28 0047c62
Add draft metadata crosswalk for CFF
