### 2025-09-05 b3c5aa8
Bump bagit version to 1.9
