+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+### Changed
+- Version of dependency bagit updated from ~=1.7 to ~=1.9 (deprecated use of pkg_resources)
+
 ## v3.3.0
 
 ### Added
