+++ pyproject.toml
@@ -32,7 +32,7 @@ classifiers = [
     'Programming Language :: Python :: 3.13',
 ]
 dependencies = [
-    "bagit~=1.7",
+    "bagit~=1.9",
     "pypandoc~=1.6",
     "python-dotenv~=1.1",
     "python-frontmatter~=1.1",
