### 2025-09-22 ea1ae65
Add option to keep previous version DOI in Codemeta file
