### 2025-09-24 #7330 @atrisovic
Hi @danielskatz, thank you! I didn’t realize the ball was in my court again. I’ll go ahead and do the recommend-accept now. And yes, I’ll make a point to check my JOSS queue more regularly now that people are back from vacation.
