### 2025-09-24 #7330 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
