### 2025-09-25 #7330 @danielskatz
An alternative could be "FACILE-RS: making it easy to archive and preserve research software repositories"
