### 2025-09-30 #7330 @danielskatz
@MarieHouillon - I was going to suggest some wording changes via a PR. In order to do so, I created an account on the GitLab instance using my GitHub credentials, which worked fine, but I cannot fork the repo, as shown here.  JOSS publication does require that anyone can create issues and PRs. How do you suggest we proceed?
