### 2025-09-30 #7330 @MarieHouillon
PRs can be created after asking for "developer" role in the project. Issues can be created by anyone.
Is this sufficient?
