+++ docs/joss/paper.bib
@@ -150,7 +150,7 @@ DOI = {10.3390/ijgi5030028}
     doi =       {10.17487/RFC8493},
     url =       {https://www.rfc-editor.org/info/rfc8493},
     author =    {John A. Kunze and Justin Littman and Liz Madden and John Scancella and Chris Adams},
-    title =     {{The BagIt File Packaging Format (V1.0)}},
+    title =     {The {BagIt} File Packaging Format ({V}1.0)},
     pagetotal = 25,
     year =      2018,
     month =     oct,
