### 2025-10-01 #7330 @danielskatz
@MarieHouillon - in the new GitHub repo, you might want to prominently say that it is a mirror intended to make it easy for people to contribute.
