### 2025-10-01 111e1c8
Update paper.bib
