+++ docs/joss/paper.bib
@@ -53,7 +53,7 @@ Publication and Citation of Research Data and Other Research Outputs. Version 4.
 
 @article{Boettiger2017,
 author = "Carl Boettiger",
-title = "{Codemeta: A Rosetta Stone for Software Metadata}",
+title = {Codemeta: A {R}osetta Stone for Software Metadata},
 year = "2017",
 month = "1",
 url = "https://figshare.com/articles/poster/Codemeta_A_Rosetta_Stone_for_Software_Metadata/4490588",
@@ -173,7 +173,7 @@ DOI = {10.3390/ijgi5030028}
   url = {https://github.com/datacite/bolognese},
   author = {Fenner, Martin},
   keywords = {doi, metadata, crossref, datacite, schema.org, bibtex, codemeta},
-  title = {Bolognese: a Ruby library for conversion of DOI Metadata},
+  title = {Bolognese: a {R}uby library for conversion of DOI Metadata},
   publisher = {DataCite},
   note = {Ruby gem and command-line utility for conversion of DOI metadata from and to different metadata formats, including schema.org.},
   year = {2017}
