### 2025-10-01 #7330 @danielskatz
I also noticed two small bib fixes I missed previously, so am trying a PR in the GitHub repo to indicate than and also see how the mirroring works - See https://github.com/openCARP-org/FACILE-RS/pull/1
