### 2025-10-01 #7330 @danielskatz
Congratulations to @MarieHouillon (Marie Houillon) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers, and we couldn't do this without you
